refactor: change structure for uncontrolled form
