--- src/components/UncontrolledForm/UncontrolledForm.type.ts
@@ -1,3 +1,0 @@
-export type FormErrors = {
-  [K in string]: { errors: string[] };
-};
