+++ src/components/index.ts
@@ -1,3 +1,2 @@
-export { UncontrolledForm } from './UncontrolledForm/UncontrolledForm';
 export { DataFormList } from './DataFormList/DataFormList';
 export { DataForm } from './DataForm/DataForm';
