+++ src/features/uncontrolledForm/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useUncontrolledForm } from './useUncontrolledForm';
