+++ src/features/uncontrolledForm/index.ts
@@ -1,0 +1,1 @@
+export { UncontrolledForm } from './components/UncontrolledForm';
