+++ src/features/uncontrolledForm/types/formErrors.ts
@@ -1,0 +1,5 @@
+import type { FormFields } from '../../../schemas';
+
+export type FormErrors = {
+  [K in keyof FormFields]: { errors: string[] };
+};
