+++ src/features/uncontrolledForm/types/index.ts
@@ -1,0 +1,1 @@
+export { type FormErrors } from './formErrors';
